Rename index -> FlexibleLineEditor
